+++ test/Main.hs
@@ -13,6 +13,7 @@ import Pudu.Frontend.ParserStateNameSpec (parserStateNameProperties)
 import Pudu.Frontend.ParserImportSpec (parserImportProperties)
 import Pudu.Frontend.ParserBindingSpec (parserBindingProperties)
 import Pudu.Frontend.ParserBlockSpec (parserBlockProperties)
+import Pudu.Frontend.ParserFunctionSpec (parserFunctionProperties)
 import Pudu.Frontend.ParserExpressionSpec (parserExpressionProperties)
 import Pudu.Frontend.ParserTypeSpec (parserTypeProperties)
 import Pudu.Frontend.SyntaxSpec (syntaxProperties)
@@ -51,9 +52,10 @@ main = do
   parserImportOutcomes <- traverse (uncurry check) parserImportProperties
   parserBindingOutcomes <- traverse (uncurry check) parserBindingProperties
   parserBlockOutcomes <- traverse (uncurry check) parserBlockProperties
+  parserFunctionOutcomes <- traverse (uncurry check) parserFunctionProperties
   parserTypeOutcomes <- traverse (uncurry check) parserTypeProperties
   parserExpressionOutcomes <- traverse (uncurry check) parserExpressionProperties
-  unless (and (sourceOutcomes <> diagnosticOutcomes <> tokenOutcomes <> cursorOutcomes <> scannerOutcomes <> numberSymbolOutcomes <> quotedOutcomes <> lexerOutcomes <> syntaxOutcomes <> parserStateNameOutcomes <> parserImportOutcomes <> parserBindingOutcomes <> parserBlockOutcomes <> parserTypeOutcomes <> parserExpressionOutcomes)) exitFailure
+  unless (and (sourceOutcomes <> diagnosticOutcomes <> tokenOutcomes <> cursorOutcomes <> scannerOutcomes <> numberSymbolOutcomes <> quotedOutcomes <> lexerOutcomes <> syntaxOutcomes <> parserStateNameOutcomes <> parserImportOutcomes <> parserBindingOutcomes <> parserBlockOutcomes <> parserFunctionOutcomes <> parserTypeOutcomes <> parserExpressionOutcomes)) exitFailure
 check :: String -> IO Property -> IO Bool
 check label loadProperty = do
   putStrLn ("[test] " <> label)